Kafka streams vs kSQL DB
Understand how filterting can be effectively applied to events
published on Kafka topcis.

We have two approaches,
1. Kafka Streams
2. kSQL DB

While kSQL DB gives a declarative way of filtering the events from
the in topics and publishes to out topics, it can only work on event
payload. Kafka streams can work on payload as well as headers.

Headers is a way to make sure filtering can be effectively applied
without going through the payload provided there are standards followed
in defining it.

This is a small study project for this analysis.